move testgrid apis to api
